multiprintscreen/requirements.txt
pillow>=10.0.0
pynput>=1.7.6
pystray>=0.19.5

